Update e2e tests for partial book drafting being enabled for everyone
