+++ src/SIL.XForge.Scripture/ClientApp/e2e/test_characterization.json
@@ -8,7 +8,7 @@
     "failure": 0
   },
   "generate_draft": {
-    "success": 13,
+    "success": 7,
     "failure": 0
   },
   "community_checking": {
@@ -16,7 +16,7 @@
     "failure": 0
   },
   "localized_screenshots": {
-    "success": 13,
+    "success": 7,
     "failure": 0
   },
   "edit_translation": {
@@ -24,10 +24,6 @@
     "failure": 3
   },
   "onboarding_flow": {
-    "success": 7,
-    "failure": 0
-  },
-  "partial_book_drafting": {
     "success": 7,
     "failure": 0
   }
